move user struct and the function receivers to separate file
